requirements.txt
asgiref==3.4.1
Django==3.2.6
django-ckeditor==6.1.0
django-js-asset==1.2.2
pkg-resources==0.0.0
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
